Arm, L2C-310: Don't separate clean and invalidates
Errata 727915 applies for PL310 revisions r2p0, r3p0, fixed in r3p1.
i.MX6 has r3p1, so is not affected. Unknown what revision exynos4
and zynq7000 have.

If a new write arrives during the clean & invalidate operation,
there is a small chance it will get lost.

The same is true when doing separate clean and invalidate operations,
hence that is not a proper work-around and worse than the errata.

The errata only applies to clean and invalidate by way, not by PA.

Signed-off-by: Indan Zupancic <indan@nul.nu>