boot: check for size_bits=0 before BIT(size_bits)
This would previously create regions that were off-by-1 on the upper
end with addresses such as 0x40001. This seems to have been fine, and
just created an extra unnecessary memory region covering the extra
bootinfo memory. But at the same time this was inconsistent with e.g.
calculate_rootserver_size() which contained the check:

  size += extra_bi_size_bits > 0 ? BIT(extra_bi_size_bits) : 0;

and this created issues when I was experimenting with some boot code
changes as the bootinfo size accounting was inconsistent.

Signed-off-by: julia <git.ts@trainwit.ch>